Extract more rustdoc-specific code
