+++ crates/indexed_docs/src/indexed_docs.rs
@@ -3,7 +3,6 @@ mod providers;
 mod registry;
 mod store;
 
-pub use crate::indexer::{DocsDotRsProvider, LocalProvider, RustdocSource};
 pub use crate::providers::rustdoc::*;
 pub use crate::registry::*;
 pub use crate::store::*;
